+++ lib/src/time_utils.dart
@@ -1,5 +1,0 @@
-class TimeUtils {
-  static final Stopwatch _stopwatch = Stopwatch()..start();
-
-  static int get elapsedMicroseconds => _stopwatch.elapsedMicroseconds;
-}