+++ lib/synchronization/condition_variable.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 
 import '../src/synchronization/wait_queue.dart';
-import '../src/time_utils.dart';
 import 'lock.dart';
 
 /// A [ConditionVariable] is a synchronization primitive  that allows to wait
@@ -11,6 +10,8 @@ import 'lock.dart';
 /// It is always used in conjunction with a locking to safely manage access to
 /// the shared data and prevent race conditions.
 class ConditionVariable {
+  static final Stopwatch _stopwatch = Stopwatch()..start();
+
   static final Future<void> _void = Future.value();
 
   final Lock lock;
@@ -50,7 +51,7 @@ class ConditionVariable {
   /// Returns `true` if the timeout has not expired; otherwise, returns `false`.
   @useResult
   Future<bool> tryWait(Duration timeout) async {
-    final started = TimeUtils.elapsedMicroseconds;
+    final started = _stopwatch.elapsedMicroseconds;
     if (timeout.isNegative) {
       throw ArgumentError.value(
           timeout, 'timeout', 'Timeout must not be negative');
@@ -59,7 +60,7 @@ class ConditionVariable {
     await lock.release();
     await _queue.enqueue();
     await lock.reacquire();
-    return TimeUtils.elapsedMicroseconds - started <= timeout.inMicroseconds;
+    return _stopwatch.elapsedMicroseconds - started <= timeout.inMicroseconds;
   }
 
   /// Releases the lock and waits for a notification.\